Seperate audio/video dowloading & merge them with ffmpeg
As the quality is set to "highestvideo", the downloaded media files often
do not contain audio in it. In some cases, we have to download the audio
part individually as well.

Seperating the downloading for video part and audio part makes sense,
as we cannot verify whether the media file actually has audio. Once the
audio part and video part are both downloaded, they are merged using
ffmpeg.